feat(ibl/shaders): add Cook-Torrance GGX prefiltered specular map with tunable LOD bias
- Add prefilter_ggx_frag shader & PrefilterGGXMaterial
- Expose Texture::setMipmapRange

There are two key improvements to the LearnOpenGL model.
1. Artifacts/banding/morphing increase with roughness due to strong
   pin-point lights and very low mip levels (1x1, 2x2). ThreeJS address
this by using a 2D atlas and generating multiple LOD 4 (16x16) mips. I
chose to implement this by ignoring mips < 16x16.

2. The "fix" suggested by LearnOpenGL for "bright dots" is to not
   directly sample the environment. The method shown still caused bright
dots on some HDRs. I added a tunable "prefilterLodBias" that allows the
user to bias towards blurrier mip levels which will soften lights and
prevent these artifacts. A value of 2.0 seems to be stable across a
variety of HDRs, but you may want to lower for accuracy, especially in
indoor HDRs.